refactor(synth-md): split tokenizer into cohesive modules
The single-pass markdown tokenizer in `packages/synth-md/src/tokenizer.ts`
(~1376 LOC) bundled the dispatch loop together with every block-level
handler. Every handler was a `private` method that touched no instance
state (the `Tokenizer` class held no fields), so they were already pure
functions of their arguments.

Extract those handlers into cohesive sub-modules under `./tokenizer/`,
behind the unchanged public `Tokenizer` surface:

- scanners.ts     - isLineWhitespace / getIndentLevel
- leaf-blocks.ts  - heading, list item, hr, blockquote, link reference,
                    paragraph, blank line
- code-blocks.ts  - fenced + indented code blocks
- setext.ts       - setext (underlined) headings
- tables.ts       - GFM tables (+ row / alignment parsing)
- html-blocks.ts  - CommonMark HTML block types 1-7

`tokenizer.ts` keeps the `Tokenizer` class, `createTokenizer()`, and the
single-pass dispatch loop, which now delegates to the extracted functions.

Pure mechanical code-motion, zero logic change. The public API is
identical: the emitted declaration signatures across the whole package
are byte-for-byte unchanged (only `private` member placeholders drop out
of the `.d.ts`, which are not externally reachable).

Gates (package @sylphx/synth-md), green at baseline and after:
- tsc --noEmit       : pass
- biome check src    : pass
- bun test           : 341 pass / 0 fail (11 files)
