+++ packages/synth-md/src/tokenizer.ts
@@ -8,22 +8,27 @@
  * 4. Minimal allocations
  *
  * Target: 2-3x improvement over OptimizedTokenizer
+ *
+ * The block-level handlers are organized into cohesive sub-modules under
+ * `./tokenizer/`; this file owns the public `Tokenizer` surface and the
+ * single-pass dispatch loop that delegates to those handlers.
  */
 
-import type {
-  BlankLineToken,
-  BlockquoteToken,
-  BlockToken,
-  CodeBlockToken,
-  HeadingToken,
-  HorizontalRuleToken,
-  HTMLBlockToken,
-  LinkReferenceToken,
-  ListItemToken,
-  ParagraphToken,
-  TableToken,
-} from './tokens.js'
-import { createPosition, createTokenPosition } from './tokens.js'
+import { parseCodeBlock, tryIndentedCodeBlock } from './tokenizer/code-blocks.js'
+import { tryHTMLBlock } from './tokenizer/html-blocks.js'
+import {
+  createBlankLine,
+  createParagraph,
+  tryBlockquote,
+  tryHeading,
+  tryHorizontalRule,
+  tryLinkReference,
+  tryListItem,
+} from './tokenizer/leaf-blocks.js'
+import { isLineWhitespace } from './tokenizer/scanners.js'
+import { trySetextHeading } from './tokenizer/setext.js'
+import { tryTable } from './tokenizer/tables.js'
+import type { BlockToken } from './tokens.js'
 
 /**
  * Ultra-Optimized Tokenizer
@@ -53,8 +58,8 @@ export class Tokenizer {
       const lineLength = lineEnd - lineStart
 
       // Fast blank line check (no substring allocation)
-      if (lineLength === 0 || this.isLineWhitespace(text, lineStart, lineEnd)) {
-        tokens.push(this.createBlankLine(text, lineStart, lineEnd, lineIndex, offset))
+      if (lineLength === 0 || isLineWhitespace(text, lineStart, lineEnd)) {
+        tokens.push(createBlankLine(text, lineStart, lineEnd, lineIndex, offset))
         offset = lineEnd + 1
         lineIndex++
         continue
@@ -64,14 +69,7 @@ export class Tokenizer {
 
       // Indented code block detection (4 spaces or tab)
       if (firstChar === ' ' || firstChar === '\t') {
-        const result = this.tryIndentedCodeBlock(
-          text,
-          lineStart,
-          lineEnd,
-          lineIndex,
-          offset,
-          length
-        )
+        const result = tryIndentedCodeBlock(text, lineStart, lineEnd, lineIndex, offset, length)
         if (result) {
           tokens.push(result.token)
           offset = result.nextOffset
@@ -87,7 +85,7 @@ export class Tokenizer {
         text[lineStart + 1] === '`' &&
         text[lineStart + 2] === '`'
       ) {
-        const result = this.parseCodeBlock(text, lineStart, lineEnd, lineIndex, length)
+        const result = parseCodeBlock(text, lineStart, lineEnd, lineIndex, length)
         if (result) {
           tokens.push(result.token)
           offset = result.nextOffset
@@ -98,7 +96,7 @@ export class Tokenizer {
 
       // Heading detection (character-based)
       if (firstChar === '#') {
-        const token = this.tryHeading(text, lineStart, lineEnd, lineIndex, offset)
+        const token = tryHeading(text, lineStart, lineEnd, lineIndex, offset)
         if (token) {
           tokens.push(token)
           offset = lineEnd + 1
@@ -109,7 +107,7 @@ export class Tokenizer {
 
       // Horizontal rule detection (must come before list items)
       if (firstChar === '-' || firstChar === '*' || firstChar === '_') {
-        const hrToken = this.tryHorizontalRule(text, lineStart, lineEnd, lineIndex, offset)
+        const hrToken = tryHorizontalRule(text, lineStart, lineEnd, lineIndex, offset)
         if (hrToken) {
           tokens.push(hrToken)
           offset = lineEnd + 1
@@ -125,7 +123,7 @@ export class Tokenizer {
         firstChar === '+' ||
         (firstChar >= '0' && firstChar <= '9')
       ) {
-        const token = this.tryListItem(text, lineStart, lineEnd, lineIndex, offset)
+        const token = tryListItem(text, lineStart, lineEnd, lineIndex, offset)
         if (token) {
           tokens.push(token)
           offset = lineEnd + 1
@@ -136,7 +134,7 @@ export class Tokenizer {
 
       // Blockquote detection
       if (firstChar === '>') {
-        const token = this.tryBlockquote(text, lineStart, lineEnd, lineIndex, offset)
+        const token = tryBlockquote(text, lineStart, lineEnd, lineIndex, offset)
         if (token) {
           tokens.push(token)
           offset = lineEnd + 1
@@ -147,7 +145,7 @@ export class Tokenizer {
 
       // GFM Table detection
       if (firstChar === '|') {
-        const result = this.tryTable(text, lineStart, lineEnd, lineIndex, offset, length)
+        const result = tryTable(text, lineStart, lineEnd, lineIndex, offset, length)
         if (result) {
           tokens.push(result.token)
           offset = result.nextOffset
@@ -158,7 +156,7 @@ export class Tokenizer {
 
       // HTML block detection
       if (firstChar === '<') {
-        const result = this.tryHTMLBlock(text, lineStart, lineEnd, lineIndex, offset, length)
+        const result = tryHTMLBlock(text, lineStart, lineEnd, lineIndex, offset, length)
         if (result) {
           tokens.push(result.token)
           offset = result.nextOffset
@@ -169,7 +167,7 @@ export class Tokenizer {
 
       // Link reference definition detection
       if (firstChar === '[') {
-        const token = this.tryLinkReference(text, lineStart, lineEnd, lineIndex, offset)
+        const token = tryLinkReference(text, lineStart, lineEnd, lineIndex, offset)
         if (token) {
           tokens.push(token)
           offset = lineEnd + 1
@@ -179,14 +177,7 @@ export class Tokenizer {
       }
 
       // Setext heading detection (lookahead to next line)
-      const setextResult = this.trySetextHeading(
-        text,
-        lineStart,
-        lineEnd,
-        lineIndex,
-        offset,
-        length
-      )
+      const setextResult = trySetextHeading(text, lineStart, lineEnd, lineIndex, offset, length)
       if (setextResult) {
         tokens.push(setextResult.token)
         offset = setextResult.nextOffset
@@ -195,1176 +186,12 @@ export class Tokenizer {
       }
 
       // Default: paragraph
-      tokens.push(this.createParagraph(text, lineStart, lineEnd, lineIndex, offset))
+      tokens.push(createParagraph(text, lineStart, lineEnd, lineIndex, offset))
       offset = lineEnd + 1
       lineIndex++
     }
 
     return tokens
-  }
-
-  /**
-   * Check if line is whitespace without substring allocation
-   */
-  private isLineWhitespace(text: string, start: number, end: number): boolean {
-    for (let i = start; i < end; i++) {
-      const c = text[i]!
-      if (c !== ' ' && c !== '\t' && c !== '\r') {
-        return false
-      }
-    }
-    return true
-  }
-
-  /**
-   * Character-based heading detection (no regex)
-   */
-  private tryHeading(
-    text: string,
-    lineStart: number,
-    lineEnd: number,
-    lineIndex: number,
-    offset: number
-  ): HeadingToken | null {
-    let depth = 0
-    let i = lineStart
-
-    // Count # symbols
-    while (i < lineEnd && text[i] === '#' && depth < 6) {
-      depth++
-      i++
-    }
-
-    if (depth === 0 || depth > 6) return null
-
-    // Must have space after #
-    if (i >= lineEnd || text[i] !== ' ') return null
-
-    i++ // Skip space
-
-    // Extract heading text (no substring until needed)
-    if (i >= lineEnd) return null
-
-    const headingText = text.slice(i, lineEnd)
-    const raw = text.slice(lineStart, lineEnd)
-
-    return {
-      type: 'heading',
-      depth: depth as 1 | 2 | 3 | 4 | 5 | 6,
-      text: headingText,
-      raw,
-      position: createTokenPosition(
-        createPosition(lineIndex, 0, offset),
-        createPosition(lineIndex, lineEnd - lineStart, offset + (lineEnd - lineStart))
-      ),
-    }
-  }
-
-  /**
-   * Character-based list item detection (no regex)
-   */
-  private tryListItem(
-    text: string,
-    lineStart: number,
-    lineEnd: number,
-    lineIndex: number,
-    offset: number
-  ): ListItemToken | null {
-    let i = lineStart
-
-    // Count leading spaces
-    let indent = 0
-    while (i < lineEnd && (text[i] === ' ' || text[i] === '\t')) {
-      indent++
-      i++
-    }
-
-    if (i >= lineEnd) return null
-
-    const markerChar = text[i]!
-
-    // Check for bullet markers (-, *, +)
-    if (markerChar === '-' || markerChar === '*' || markerChar === '+') {
-      i++
-    }
-    // Check for numbered list (1., 2., etc.)
-    else if (markerChar >= '0' && markerChar <= '9') {
-      // Skip digits
-      while (i < lineEnd && text[i]! >= '0' && text[i]! <= '9') {
-        i++
-      }
-      // Must have .
-      if (i >= lineEnd || text[i] !== '.') return null
-      i++ // Skip .
-    } else {
-      return null
-    }
-
-    // Must have space after marker
-    if (i >= lineEnd || text[i] !== ' ') return null
-    i++ // Skip space
-
-    // Extract text
-    const itemText = i < lineEnd ? text.slice(i, lineEnd) : ''
-
-    // Check for task list
-    let checked: boolean | undefined
-    if (itemText.startsWith('[') && itemText[2] === ']') {
-      const checkChar = itemText[1]
-      if (checkChar === 'x' || checkChar === ' ') {
-        checked = checkChar === 'x'
-      }
-    }
-
-    const raw = text.slice(lineStart, lineEnd)
-
-    return {
-      type: 'listItem',
-      indent,
-      text: itemText,
-      checked,
-      raw,
-      position: createTokenPosition(
-        createPosition(lineIndex, 0, offset),
-        createPosition(lineIndex, lineEnd - lineStart, offset + (lineEnd - lineStart))
-      ),
-    }
-  }
-
-  /**
-   * Character-based horizontal rule detection
-   */
-  private tryHorizontalRule(
-    text: string,
-    lineStart: number,
-    lineEnd: number,
-    lineIndex: number,
-    offset: number
-  ): HorizontalRuleToken | null {
-    const firstChar = text[lineStart]!
-    if (firstChar !== '-' && firstChar !== '*' && firstChar !== '_') {
-      return null
-    }
-
-    let count = 0
-    for (let i = lineStart; i < lineEnd; i++) {
-      const c = text[i]!
-      if (c === firstChar) {
-        count++
-      } else if (c !== ' ' && c !== '\t') {
-        return null
-      }
-    }
-
-    if (count < 3) return null
-
-    const raw = text.slice(lineStart, lineEnd)
-
-    return {
-      type: 'horizontalRule',
-      raw,
-      position: createTokenPosition(
-        createPosition(lineIndex, 0, offset),
-        createPosition(lineIndex, lineEnd - lineStart, offset + (lineEnd - lineStart))
-      ),
-    }
-  }
-
-  /**
-   * Character-based blockquote detection
-   */
-  private tryBlockquote(
-    text: string,
-    lineStart: number,
-    lineEnd: number,
-    lineIndex: number,
-    offset: number
-  ): BlockquoteToken | null {
-    if (text[lineStart] !== '>') return null
-
-    let i = lineStart + 1
-
-    // Skip optional space
-    if (i < lineEnd && text[i] === ' ') {
-      i++
-    }
-
-    const quoteText = i < lineEnd ? text.slice(i, lineEnd) : ''
-    const raw = text.slice(lineStart, lineEnd)
-
-    return {
-      type: 'blockquote',
-      text: quoteText,
-      raw,
-      position: createTokenPosition(
-        createPosition(lineIndex, 0, offset),
-        createPosition(lineIndex, lineEnd - lineStart, offset + (lineEnd - lineStart))
-      ),
-    }
-  }
-
-  /**
-   * Parse code block without split()
-   */
-  private parseCodeBlock(
-    text: string,
-    lineStart: number,
-    lineEnd: number,
-    startLine: number,
-    textLength: number
-  ): { token: CodeBlockToken; nextOffset: number; nextLine: number } | null {
-    // Parse opening fence
-    let i = lineStart + 3 // Skip ```
-
-    // Skip spaces
-    while (i < lineEnd && text[i] === ' ') {
-      i++
-    }
-
-    // Extract language
-    const langStart = i
-    while (i < lineEnd && text[i] !== ' ' && text[i] !== '\n') {
-      i++
-    }
-    const lang = langStart < i ? text.slice(langStart, i) : undefined
-
-    // Extract meta
-    while (i < lineEnd && text[i] === ' ') {
-      i++
-    }
-    const meta = i < lineEnd ? text.slice(i, lineEnd) : undefined
-
-    // Find code content and closing fence
-    let offset = lineEnd + 1
-    let currentLine = startLine + 1
-    const codeStart = offset
-
-    while (offset < textLength) {
-      const fenceLineStart = offset
-
-      // Find end of line
-      while (offset < textLength && text[offset] !== '\n') {
-        offset++
-      }
-
-      // Check for closing fence
-      if (
-        offset - fenceLineStart >= 3 &&
-        text[fenceLineStart] === '`' &&
-        text[fenceLineStart + 1] === '`' &&
-        text[fenceLineStart + 2] === '`'
-      ) {
-        // Found closing fence
-        const code = codeStart < fenceLineStart - 1 ? text.slice(codeStart, fenceLineStart - 1) : ''
-
-        return {
-          token: {
-            type: 'codeBlock',
-            lang,
-            meta,
-            code,
-            raw: text.slice(lineStart, offset),
-            position: createTokenPosition(
-              createPosition(startLine, 0, lineStart),
-              createPosition(currentLine, offset - fenceLineStart, offset)
-            ),
-          },
-          nextOffset: offset + 1,
-          nextLine: currentLine + 1,
-        }
-      }
-
-      offset++
-      currentLine++
-    }
-
-    // No closing fence
-    const code = text.slice(codeStart)
-
-    return {
-      token: {
-        type: 'codeBlock',
-        lang,
-        meta,
-        code,
-        raw: text.slice(lineStart),
-        position: createTokenPosition(
-          createPosition(startLine, 0, lineStart),
-          createPosition(currentLine, 0, textLength)
-        ),
-      },
-      nextOffset: textLength,
-      nextLine: currentLine,
-    }
-  }
-
-  /**
-   * Create paragraph token
-   */
-  private createParagraph(
-    text: string,
-    lineStart: number,
-    lineEnd: number,
-    lineIndex: number,
-    offset: number
-  ): ParagraphToken {
-    const paragraphText = text.slice(lineStart, lineEnd)
-
-    return {
-      type: 'paragraph',
-      text: paragraphText,
-      raw: paragraphText,
-      position: createTokenPosition(
-        createPosition(lineIndex, 0, offset),
-        createPosition(lineIndex, lineEnd - lineStart, offset + (lineEnd - lineStart))
-      ),
-    }
-  }
-
-  /**
-   * Create blank line token
-   */
-  private createBlankLine(
-    text: string,
-    lineStart: number,
-    lineEnd: number,
-    lineIndex: number,
-    offset: number
-  ): BlankLineToken {
-    const raw = text.slice(lineStart, lineEnd)
-
-    return {
-      type: 'blankLine',
-      raw,
-      position: createTokenPosition(
-        createPosition(lineIndex, 0, offset),
-        createPosition(lineIndex, lineEnd - lineStart, offset + (lineEnd - lineStart))
-      ),
-    }
-  }
-
-  /**
-   * Try to parse indented code block (4 spaces or 1 tab)
-   */
-  private tryIndentedCodeBlock(
-    text: string,
-    lineStart: number,
-    lineEnd: number,
-    startLine: number,
-    startOffset: number,
-    textLength: number
-  ): { token: CodeBlockToken; nextOffset: number; nextLine: number } | null {
-    // Check if line has proper indentation (4 spaces or 1 tab)
-    const firstLineIndent = this.getIndentLevel(text, lineStart, lineEnd)
-    if (firstLineIndent < 4) return null
-
-    // Collect code lines
-    const codeLines: string[] = []
-    let currentOffset = lineStart
-    let currentLine = startLine
-
-    while (currentOffset < textLength) {
-      const rowStart = currentOffset
-      let rowEnd = currentOffset
-
-      // Find end of line
-      while (rowEnd < textLength && text[rowEnd] !== '\n') {
-        rowEnd++
-      }
-
-      const lineIndent = this.getIndentLevel(text, rowStart, rowEnd)
-
-      // Empty line - include it and continue
-      if (rowStart === rowEnd || this.isLineWhitespace(text, rowStart, rowEnd)) {
-        codeLines.push('')
-        currentOffset = rowEnd + 1
-        currentLine++
-        continue
-      }
-
-      // Indented line - include it
-      if (lineIndent >= 4) {
-        // Remove 4 spaces of indentation
-        let codeStart = rowStart
-        let removed = 0
-        while (removed < 4 && codeStart < rowEnd) {
-          if (text[codeStart] === '\t') {
-            removed = 4
-            codeStart++
-          } else if (text[codeStart] === ' ') {
-            removed++
-            codeStart++
-          } else {
-            break
-          }
-        }
-
-        const codeLine = text.slice(codeStart, rowEnd)
-        codeLines.push(codeLine)
-        currentOffset = rowEnd + 1
-        currentLine++
-        continue
-      }
-
-      // Non-indented line, end of code block
-      break
-    }
-
-    // Remove trailing blank lines
-    while (codeLines.length > 0 && codeLines[codeLines.length - 1] === '') {
-      codeLines.pop()
-      currentLine--
-    }
-
-    if (codeLines.length === 0) {
-      return null
-    }
-
-    const code = codeLines.join('\n')
-    const endOffset = currentOffset > 0 ? currentOffset - 1 : currentOffset
-    const raw = text.slice(lineStart, endOffset)
-
-    return {
-      token: {
-        type: 'codeBlock',
-        lang: undefined,
-        meta: undefined,
-        code,
-        raw,
-        position: createTokenPosition(
-          createPosition(startLine, 0, startOffset),
-          createPosition(currentLine - 1, 0, endOffset)
-        ),
-      },
-      nextOffset: currentOffset,
-      nextLine: currentLine,
-    }
-  }
-
-  /**
-   * Get indentation level of a line
-   */
-  private getIndentLevel(text: string, lineStart: number, lineEnd: number): number {
-    let indent = 0
-    let i = lineStart
-
-    while (i < lineEnd && (text[i] === ' ' || text[i] === '\t')) {
-      if (text[i] === '\t') {
-        return 4 // Tab counts as 4 spaces
-      }
-      indent++
-      i++
-    }
-
-    return indent
-  }
-
-  /**
-   * Try to parse Setext heading (with underline)
-   */
-  private trySetextHeading(
-    text: string,
-    lineStart: number,
-    lineEnd: number,
-    startLine: number,
-    startOffset: number,
-    textLength: number
-  ): { token: HeadingToken; nextOffset: number; nextLine: number } | null {
-    // Need at least one more line
-    if (lineEnd + 1 >= textLength) return null
-
-    // Find next line boundaries
-    const nextLineStart = lineEnd + 1
-    let nextLineEnd = nextLineStart
-
-    while (nextLineEnd < textLength && text[nextLineEnd] !== '\n') {
-      nextLineEnd++
-    }
-
-    // Check if next line is a setext underline (=== or ---)
-    if (nextLineEnd <= nextLineStart) return null
-
-    const underlineChar = text[nextLineStart]!
-
-    // Must be = or -
-    if (underlineChar !== '=' && underlineChar !== '-') return null
-
-    // Check if entire line is the underline character (with optional spaces)
-    let validUnderline = true
-    for (let i = nextLineStart; i < nextLineEnd; i++) {
-      const c = text[i]!
-      if (c !== underlineChar && c !== ' ' && c !== '\t') {
-        validUnderline = false
-        break
-      }
-    }
-
-    if (!validUnderline) return null
-
-    // Valid setext heading
-    const depth = underlineChar === '=' ? 1 : 2
-    const headingText = text.slice(lineStart, lineEnd)
-    const raw = text.slice(lineStart, nextLineEnd)
-
-    return {
-      token: {
-        type: 'heading',
-        depth: depth as 1 | 2,
-        text: headingText,
-        raw,
-        position: createTokenPosition(
-          createPosition(startLine, 0, startOffset),
-          createPosition(startLine + 1, nextLineEnd - nextLineStart, nextLineEnd + startOffset)
-        ),
-      },
-      nextOffset: nextLineEnd + 1,
-      nextLine: startLine + 2,
-    }
-  }
-
-  /**
-   * Try to parse GFM table with lookahead
-   */
-  private tryTable(
-    text: string,
-    headerStart: number,
-    headerEnd: number,
-    startLine: number,
-    startOffset: number,
-    textLength: number
-  ): { token: TableToken; nextOffset: number; nextLine: number } | null {
-    // Parse header row
-    const headerRow = this.parseTableRow(text, headerStart, headerEnd)
-    if (!headerRow || headerRow.length === 0) return null
-
-    // Lookahead to next line for separator
-    const separatorStart = headerEnd + 1
-    if (separatorStart >= textLength) return null
-
-    // Find separator line boundaries
-    let separatorEnd = separatorStart
-    while (separatorEnd < textLength && text[separatorEnd] !== '\n') {
-      separatorEnd++
-    }
-
-    // Parse and validate separator
-    const align = this.parseTableAlignment(text, separatorStart, separatorEnd)
-    if (!align || align.length !== headerRow.length) return null
-
-    // Parse data rows
-    const rows: string[][] = []
-    let currentOffset = separatorEnd + 1
-    let currentLine = startLine + 2
-
-    while (currentOffset < textLength) {
-      const rowStart = currentOffset
-      let rowEnd = currentOffset
-
-      // Find end of line
-      while (rowEnd < textLength && text[rowEnd] !== '\n') {
-        rowEnd++
-      }
-
-      // Check if line is part of table
-      if (rowEnd > rowStart && text[rowStart] === '|') {
-        const row = this.parseTableRow(text, rowStart, rowEnd)
-        if (row && row.length > 0) {
-          rows.push(row)
-          currentOffset = rowEnd + 1
-          currentLine++
-          continue
-        }
-      }
-
-      // Not a table row, end of table
-      break
-    }
-
-    // Build raw content
-    const raw = text.slice(headerStart, currentOffset - 1)
-
-    return {
-      token: {
-        type: 'table',
-        header: headerRow,
-        align,
-        rows,
-        raw,
-        position: createTokenPosition(
-          createPosition(startLine, 0, startOffset),
-          createPosition(currentLine - 1, 0, currentOffset - 1)
-        ),
-      },
-      nextOffset: currentOffset,
-      nextLine: currentLine,
-    }
-  }
-
-  /**
-   * Parse a table row (| cell | cell |)
-   */
-  private parseTableRow(text: string, lineStart: number, lineEnd: number): string[] | null {
-    const cells: string[] = []
-    let i = lineStart
-
-    // Skip leading |
-    if (i < lineEnd && text[i] === '|') {
-      i++
-    }
-
-    let cellStart = i
-
-    while (i <= lineEnd) {
-      const char = i < lineEnd ? text[i] : null
-
-      if (char === '|' || char === null) {
-        // Extract cell content
-        const cellContent = text.slice(cellStart, i).trim()
-        cells.push(cellContent)
-
-        // Move to next cell
-        i++
-        cellStart = i
-      } else {
-        i++
-      }
-    }
-
-    // Remove trailing empty cell if line ends with |
-    if (cells.length > 0 && cells[cells.length - 1] === '') {
-      cells.pop()
-    }
-
-    return cells.length > 0 ? cells : null
-  }
-
-  /**
-   * Parse table alignment row (|:---|:---:|---:|)
-   */
-  private parseTableAlignment(
-    text: string,
-    lineStart: number,
-    lineEnd: number
-  ): Array<'left' | 'right' | 'center' | null> | null {
-    const cells = this.parseTableRow(text, lineStart, lineEnd)
-    if (!cells) return null
-
-    const alignments: Array<'left' | 'right' | 'center' | null> = []
-
-    for (const cell of cells) {
-      // Check if valid separator (----, :---, ---:, :---:)
-      const trimmed = cell.trim()
-
-      if (trimmed.length < 3) return null
-
-      const startsWithColon = trimmed[0] === ':'
-      const endsWithColon = trimmed[trimmed.length - 1] === ':'
-
-      // Check all middle characters are dashes
-      const start = startsWithColon ? 1 : 0
-      const end = endsWithColon ? trimmed.length - 1 : trimmed.length
-
-      for (let i = start; i < end; i++) {
-        if (trimmed[i] !== '-') return null
-      }
-
-      // Determine alignment
-      if (startsWithColon && endsWithColon) {
-        alignments.push('center')
-      } else if (startsWithColon) {
-        alignments.push('left')
-      } else if (endsWithColon) {
-        alignments.push('right')
-      } else {
-        alignments.push(null)
-      }
-    }
-
-    return alignments
-  }
-
-  /**
-   * Try to parse HTML block
-   */
-  private tryHTMLBlock(
-    text: string,
-    lineStart: number,
-    lineEnd: number,
-    startLine: number,
-    startOffset: number,
-    textLength: number
-  ): { token: HTMLBlockToken; nextOffset: number; nextLine: number } | null {
-    if (text[lineStart] !== '<') return null
-
-    // Type 1: Script/pre/style/textarea tags
-    const scriptTags = ['<script', '<pre', '<style', '<textarea']
-    for (const tag of scriptTags) {
-      if (this.startsWith(text, lineStart, lineEnd, tag)) {
-        return this.parseHTMLBlockWithClosingTag(
-          text,
-          lineStart,
-          startLine,
-          startOffset,
-          textLength,
-          tag.slice(1) // Remove <
-        )
-      }
-    }
-
-    // Type 2: HTML comment <!-- -->
-    if (this.startsWith(text, lineStart, lineEnd, '<!--')) {
-      return this.parseHTMLBlockUntil(text, lineStart, startLine, startOffset, textLength, '-->')
-    }
-
-    // Type 3: Processing instruction <? ?>
-    if (this.startsWith(text, lineStart, lineEnd, '<?')) {
-      return this.parseHTMLBlockUntil(text, lineStart, startLine, startOffset, textLength, '?>')
-    }
-
-    // Type 4: Declaration <!UPPER>
-    if (this.startsWith(text, lineStart, lineEnd, '<!') && lineStart + 2 < lineEnd) {
-      const thirdChar = text[lineStart + 2]!
-      if (thirdChar >= 'A' && thirdChar <= 'Z') {
-        return this.parseHTMLBlockUntil(text, lineStart, startLine, startOffset, textLength, '>')
-      }
-    }
-
-    // Type 5: CDATA <![CDATA[ ]]>
-    if (this.startsWith(text, lineStart, lineEnd, '<![CDATA[')) {
-      return this.parseHTMLBlockUntil(text, lineStart, startLine, startOffset, textLength, ']]>')
-    }
-
-    // Type 6: Block-level tags
-    const blockTags = [
-      'address',
-      'article',
-      'aside',
-      'base',
-      'basefont',
-      'blockquote',
-      'body',
-      'caption',
-      'center',
-      'col',
-      'colgroup',
-      'dd',
-      'details',
-      'dialog',
-      'dir',
-      'div',
-      'dl',
-      'dt',
-      'fieldset',
-      'figcaption',
-      'figure',
-      'footer',
-      'form',
-      'frame',
-      'frameset',
-      'h1',
-      'h2',
-      'h3',
-      'h4',
-      'h5',
-      'h6',
-      'head',
-      'header',
-      'hr',
-      'html',
-      'iframe',
-      'legend',
-      'li',
-      'link',
-      'main',
-      'menu',
-      'menuitem',
-      'nav',
-      'noframes',
-      'ol',
-      'optgroup',
-      'option',
-      'p',
-      'param',
-      'section',
-      'source',
-      'summary',
-      'table',
-      'tbody',
-      'td',
-      'tfoot',
-      'th',
-      'thead',
-      'title',
-      'tr',
-      'track',
-      'ul',
-    ]
-
-    for (const tag of blockTags) {
-      if (this.startsWithTag(text, lineStart, lineEnd, tag)) {
-        return this.parseHTMLBlockUntilBlankLine(
-          text,
-          lineStart,
-          startLine,
-          startOffset,
-          textLength
-        )
-      }
-    }
-
-    // Type 7: Complete tag on single line
-    if (this.isCompleteHTMLTag(text, lineStart, lineEnd)) {
-      const content = text.slice(lineStart, lineEnd)
-      const raw = content
-      return {
-        token: {
-          type: 'htmlBlock',
-          content,
-          raw,
-          position: createTokenPosition(
-            createPosition(startLine, 0, startOffset),
-            createPosition(startLine, lineEnd - lineStart, startOffset + (lineEnd - lineStart))
-          ),
-        },
-        nextOffset: lineEnd + 1,
-        nextLine: startLine + 1,
-      }
-    }
-
-    return null
-  }
-
-  /**
-   * Check if text starts with a specific string
-   */
-  private startsWith(text: string, lineStart: number, lineEnd: number, prefix: string): boolean {
-    if (lineStart + prefix.length > lineEnd) return false
-
-    for (let i = 0; i < prefix.length; i++) {
-      if (text[lineStart + i] !== prefix[i]) return false
-    }
-
-    return true
-  }
-
-  /**
-   * Check if text starts with an HTML tag
-   */
-  private startsWithTag(
-    text: string,
-    lineStart: number,
-    lineEnd: number,
-    tagName: string
-  ): boolean {
-    if (!this.startsWith(text, lineStart, lineEnd, `<${tagName}`)) return false
-
-    const afterTag = lineStart + tagName.length + 1
-    if (afterTag >= lineEnd) return true
-
-    const nextChar = text[afterTag]!
-    // Tag must be followed by space, >, or /
-    return (
-      nextChar === ' ' ||
-      nextChar === '>' ||
-      nextChar === '/' ||
-      nextChar === '\t' ||
-      nextChar === '\n'
-    )
-  }
-
-  /**
-   * Parse HTML block with specific closing tag
-   */
-  private parseHTMLBlockWithClosingTag(
-    text: string,
-    lineStart: number,
-    startLine: number,
-    startOffset: number,
-    textLength: number,
-    tagName: string
-  ): { token: HTMLBlockToken; nextOffset: number; nextLine: number } | null {
-    const closingTag = `</${tagName}>`
-    let currentOffset = lineStart
-    let currentLine = startLine
-
-    // Find closing tag
-    while (currentOffset < textLength) {
-      // Find next >
-      const gtIndex = text.indexOf('>', currentOffset)
-      if (gtIndex === -1) break
-
-      // Check if this is the closing tag
-      const possibleClosing = text.slice(Math.max(0, gtIndex - closingTag.length + 1), gtIndex + 1)
-      if (possibleClosing.toLowerCase() === closingTag.toLowerCase()) {
-        // Found closing tag
-        const endOffset = gtIndex + 1
-
-        // Count newlines to update line index
-        for (let i = lineStart; i < endOffset; i++) {
-          if (text[i] === '\n') currentLine++
-        }
-
-        const content = text.slice(lineStart, endOffset)
-        return {
-          token: {
-            type: 'htmlBlock',
-            content,
-            raw: content,
-            position: createTokenPosition(
-              createPosition(startLine, 0, startOffset),
-              createPosition(currentLine, 0, endOffset)
-            ),
-          },
-          nextOffset: endOffset + (text[endOffset] === '\n' ? 1 : 0),
-          nextLine: currentLine + (text[endOffset] === '\n' ? 1 : 0),
-        }
-      }
-
-      currentOffset = gtIndex + 1
-    }
-
-    return null
-  }
-
-  /**
-   * Parse HTML block until specific end marker
-   */
-  private parseHTMLBlockUntil(
-    text: string,
-    lineStart: number,
-    startLine: number,
-    startOffset: number,
-    textLength: number,
-    endMarker: string
-  ): { token: HTMLBlockToken; nextOffset: number; nextLine: number } | null {
-    const endIndex = text.indexOf(endMarker, lineStart)
-    if (endIndex === -1) {
-      // No end marker found, consume rest of document
-      const content = text.slice(lineStart)
-      let currentLine = startLine
-
-      for (let i = lineStart; i < textLength; i++) {
-        if (text[i] === '\n') currentLine++
-      }
-
-      return {
-        token: {
-          type: 'htmlBlock',
-          content,
-          raw: content,
-          position: createTokenPosition(
-            createPosition(startLine, 0, startOffset),
-            createPosition(currentLine, 0, textLength)
-          ),
-        },
-        nextOffset: textLength,
-        nextLine: currentLine,
-      }
-    }
-
-    const endOffset = endIndex + endMarker.length
-    let currentLine = startLine
-
-    for (let i = lineStart; i < endOffset; i++) {
-      if (text[i] === '\n') currentLine++
-    }
-
-    const content = text.slice(lineStart, endOffset)
-    return {
-      token: {
-        type: 'htmlBlock',
-        content,
-        raw: content,
-        position: createTokenPosition(
-          createPosition(startLine, 0, startOffset),
-          createPosition(currentLine, 0, endOffset)
-        ),
-      },
-      nextOffset: endOffset + (text[endOffset] === '\n' ? 1 : 0),
-      nextLine: currentLine + (text[endOffset] === '\n' ? 1 : 0),
-    }
-  }
-
-  /**
-   * Parse HTML block until blank line
-   */
-  private parseHTMLBlockUntilBlankLine(
-    text: string,
-    lineStart: number,
-    startLine: number,
-    startOffset: number,
-    textLength: number
-  ): { token: HTMLBlockToken; nextOffset: number; nextLine: number } | null {
-    let currentOffset = lineStart
-    let currentLine = startLine
-    let lastLineEnd = lineStart
-
-    while (currentOffset < textLength) {
-      const rowStart = currentOffset
-      let rowEnd = currentOffset
-
-      // Find end of line
-      while (rowEnd < textLength && text[rowEnd] !== '\n') {
-        rowEnd++
-      }
-
-      // Check if blank line
-      if (rowStart === rowEnd || this.isLineWhitespace(text, rowStart, rowEnd)) {
-        // Found blank line, end of HTML block
-        break
-      }
-
-      lastLineEnd = rowEnd
-      currentOffset = rowEnd + 1
-      currentLine++
-    }
-
-    const content = text.slice(lineStart, lastLineEnd)
-    return {
-      token: {
-        type: 'htmlBlock',
-        content,
-        raw: content,
-        position: createTokenPosition(
-          createPosition(startLine, 0, startOffset),
-          createPosition(currentLine, 0, lastLineEnd)
-        ),
-      },
-      nextOffset: currentOffset,
-      nextLine: currentLine + 1,
-    }
-  }
-
-  /**
-   * Check if line contains a complete HTML tag
-   */
-  private isCompleteHTMLTag(text: string, lineStart: number, lineEnd: number): boolean {
-    // Must start with <
-    if (text[lineStart] !== '<') return false
-
-    // Must end with >
-    let actualEnd = lineEnd - 1
-    while (actualEnd > lineStart && (text[actualEnd] === ' ' || text[actualEnd] === '\t')) {
-      actualEnd--
-    }
-
-    if (text[actualEnd] !== '>') return false
-
-    // Check for valid tag structure: <tagname> or </tagname> or <tagname />
-    let i = lineStart + 1
-    const isClosing = text[i] === '/'
-    if (isClosing) i++
-
-    // Tag name must start with letter
-    if (i >= actualEnd) return false
-    const firstChar = text[i]!
-    if (!((firstChar >= 'a' && firstChar <= 'z') || (firstChar >= 'A' && firstChar <= 'Z'))) {
-      return false
-    }
-
-    return true
-  }
-
-  /**
-   * Try to parse link reference definition: [label]: url "title"
-   */
-  private tryLinkReference(
-    text: string,
-    lineStart: number,
-    lineEnd: number,
-    lineIndex: number,
-    offset: number
-  ): LinkReferenceToken | null {
-    if (text[lineStart] !== '[') return null
-
-    let i = lineStart + 1
-
-    // Find closing ]
-    while (i < lineEnd && text[i] !== ']') {
-      i++
-    }
-
-    if (i >= lineEnd) return null
-
-    const label = text
-      .slice(lineStart + 1, i)
-      .toLowerCase()
-      .trim()
-    if (label.length === 0) return null
-
-    i++ // Skip ]
-
-    // Must have :
-    if (i >= lineEnd || text[i] !== ':') return null
-    i++ // Skip :
-
-    // Skip whitespace
-    while (i < lineEnd && (text[i] === ' ' || text[i] === '\t')) {
-      i++
-    }
-
-    if (i >= lineEnd) return null
-
-    // Parse URL
-    const urlStart = i
-
-    // URL can be in angle brackets or bare
-    let url: string
-    if (text[i] === '<') {
-      i++ // Skip <
-      const urlContentStart = i
-      while (i < lineEnd && text[i] !== '>') {
-        i++
-      }
-      if (i >= lineEnd) return null
-      url = text.slice(urlContentStart, i)
-      i++ // Skip >
-    } else {
-      // Bare URL - until whitespace or end
-      while (i < lineEnd && text[i] !== ' ' && text[i] !== '\t') {
-        i++
-      }
-      url = text.slice(urlStart, i)
-    }
-
-    if (url.length === 0) return null
-
-    // Skip whitespace
-    while (i < lineEnd && (text[i] === ' ' || text[i] === '\t')) {
-      i++
-    }
-
-    // Parse optional title
-    let title: string | undefined
-    if (i < lineEnd) {
-      const quoteChar = text[i]
-      if (quoteChar === '"' || quoteChar === "'" || quoteChar === '(') {
-        const closingQuote = quoteChar === '(' ? ')' : quoteChar
-        i++ // Skip opening quote
-        const titleStart = i
-        while (i < lineEnd && text[i] !== closingQuote) {
-          i++
-        }
-        if (i < lineEnd) {
-          title = text.slice(titleStart, i)
-          i++ // Skip closing quote
-        }
-      }
-    }
-
-    const raw = text.slice(lineStart, lineEnd)
-
-    return {
-      type: 'linkReference',
-      label,
-      url,
-      title,
-      raw,
-      position: createTokenPosition(
-        createPosition(lineIndex, 0, offset),
-        createPosition(lineIndex, lineEnd - lineStart, offset + (lineEnd - lineStart))
-      ),
-    }
   }
 }
